add skeleton for credential crud
